Because tv_check has dependencies on Tk and Tk::TableMatrix, which the other parts of xmltv don't, I have made it optional.  The Makefile.PL now asks whether you want to install tv_check.  I may do the same to other parts of XMLTV if they require non-standard modules.
This meant moving Ask.pm out of the grab/ directory (which was a
slightly silly place for it anyway) and into lib/, since it is now
used by the whole project and not just by the grabbers.  The
Makefile.PL loads this library 'by hand' and uses it to prompt about
tv_check.

I've also changed tv_check's README to standardize on the name
tv_check not TVCHECK or tvcheck... I hope it's author doesn't mind too
much :-(.