Reimplemented the parsing code in XMLTV.pm using XML::Twig instead of XML::DOM.  Thanks to Leif Van Horn for the idea, although in the end I didn't use his code.  Really all that needed changing were the wrapper routines like get_attrs() around the DOM tree methods.  The new implementations, wrapping XML::Twig, are quite a bit shorter :-).
The interface to the XMLTV module is just the same as before, reading
all the data in a big lump.  A lazy-reading interface is needed to get
the best out of XML::Twig, and I will implement that later.  But
already there is an 18% speed improvement (on the test suite) from the
old version.

I can't quite eliminate XML::DOM because it is still used by
tv_grab_uk to parse Ananova's weird 'regions' page.  But that will be
next.

Also fixed the installation of section 1 manual pages.