XMLTV.pm.in: added new cat_noprogrammes() function which combines data for encoding, credits, and channels - but not programmes.  This is in line with the general policy of allowing programmes to be processed individually rather than in one big lump.
tv_grab_uk.in: changed to use cat_noprogrammes() instead of just cat()
in an effort to reduce the working set size, or at least to give it better
locality.  Previously tv_grab_uk would rearrange the order of programmes
and then output them by channel, which thrashed the disk a lot.  This is
not the last change to tv_grab_uk to improve scalability, just an
intermediate point to commit a working version.