Adding Jerome Benedict Hettich's Radio Times grabber, which was written as part of a group project at the Department of Computing, Imperial College - another case of someone in the same department using XMLTV without me knowing!  He sent it to me a while ago but I never got around to adding it.
At the moment it does a pretty good job of fetching data from the Radio
Times site - their listings are very detailed, particularly in giving the
cast of each programme.  And they stretch to two weeks whereas Ananova's
go to one.  I believe that the RT data is 'completely unrelated' to the
Press Association feed which Ananova uses, but I'll investigate later
whether there are similarities.