Sorting is necessary to compare the output of tv_grab_uk from the 0.5 release with that in 0.5.1.  The order of the XML output has changed.
(Also it's necessary to turn off $KEEP_ENCODING in XMLTV.pm, if you
want to get the same output from the grabber.)

I'm pleased to report that with the exception of cases which crashed
the old version of the grabber, the output of old and new is unchanged
up to sorting.

BTW, if anyone else wants to know how to use this script (like to
check that tv_grab_na still works after some major change to the
code), do ask.  I plan to start using it for tv_grab_sn and
tv_grab_uk_rt as well.