For a long time now there has been a nuke_from_rel() call when a programme has been fully processed, forgetting the information of which programmes it shares a clump with.  The idea was that once a programme has been dealt with, we don't need to keep track of it any more.  'Probably not needed, but can't hurt' said the comment: tempting fate if ever that was done.
No, we may need to modify a programme even once it has been fully
munged, because if others in the clump have been deleted we will need
to update its clumpidx.  So it shouldn't be removed from the
relation.  Getting rid of that call initially caused a section of code
in Clumps.pm (which had never been run before) to break.  But it
should all be fixed now, I hope.

Also added some trace messages.