Patch from Staffen Malmgren.  Here's his description:
The following patch fixes three bugs concerning the --configure mode of
tv_grab_sn.

The first one is one i introduced myself in my patch from a couple of
months ago; the 'all' answer to "add channel FOO?" question would
incorrectly NOT add the current channel (but it would add all following
channels)

The second two were introduced in revisions 1.19, 1.20 or thereabouts
(and so does not seem to be present in 0.58). The first error was
handling a array reference as though it was scalar (and thus ending up
with a lot of "ARRAY(0xABCD1234)" in the configuration questions and
config file.  The second was a regression in the HTML scraping of
available channels -- along with the valid channels, a bunch of
categories would be returned.

The patch is tested with both Swedish and Norwegian channels.